adding migrations, adding seeds, adding images, small bugfix

+++ database/migrations/2019_11_27_151153_create_counts_table.php
@@ -19,7 +19,7 @@ class CreateCountsTable extends Migration
                 $table->string('name');
                 $table->string('type', 10);
                 $table->unsignedInteger('id_object')->nullable();
-                $table->float('impulse');
+                $table->double('impulse',10,7);
                 $table->string('unit', 6);
                 $table->float('today_value');
                 $table->float('total_value');
